pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "baserow-client"
version = "0.5.1"
description = "Client for the baserow.io API."
authors = ["Niklas Rosenstein <rosensteinniklas@gmail.com>"]
license = "MIT"
readme = "readme.md"
packages = [{ include = "baserow", from = "src" }]
classifiers = []
keywords = []

[tool.poetry.urls]
# "Bug Tracker" = ""
# Documentation = ""
Homepage = "https://github.com/NiklasRosenstein/python-baserow-client"
# Repository = ""

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
"databind.core" = "^1.2.1"
"databind.json" = "^1.2.1"
typing-extensions = ">=3.0.0"

[tool.poetry.dev-dependencies]
mypy = "*"
pytest = "*"
pycln = "^2.1.3"
types-requests = "^2.25.6"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
novella = "^0.1.14"
pydoc-markdown = "^4.6.0"
mkdocs = "*"
mkdocs-material = "*"

[tool.slap]
typed = true

[tool.slap.test]
check = "slap check"
mypy = "MYPYPATH=src mypy src/ --namespace-packages --explicit-package-bases"
pytest = "pytest src/ test/ -vv"
pycln  = "pycln src/ tests/ --check"

[tool.slap.run]
"docs:build" = "cd docs && novella --base-url python-baserow-client/"
"docs:dev" = "cd docs && novella --serve"

[tool.mypy]
pretty = true
namespace_packages = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
show_error_context = true
show_error_codes = true
